Make the port of the incident service configurable
Currently we have hard coded the port of the incident service that
we interact with to 80. However, when using internal service URLs
instead of public Routes it is more likely to be 8080.

To allow for both of these options an `incident.port` configuration
option has been added which defaults to 8080.

The default incident URL has also been made something more generic
instead of referencing a specific cluster. It still references a
specific namespace but this seems to be more generic.